% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_gwas_data}
\alias{load_gwas_data}
\title{Loading GWAS summary statistics}
\usage{
load_gwas_data(file_path, col_map = NULL)
}
\arguments{
\item{file_path}{\verb{a character string} specifying the path to the GWAS
summary statistics file (tab-limited).}

\item{col_map}{(Optional) \verb{named list} mapping required column names
(\code{SNP}, \code{beta}, \code{se}, \code{pval}, \code{effect_allele}, \code{other_allele},
\code{eaf}) to actual column names in the file.
If NULL, it assumes that the columns are already correctly named.}
}
\value{
a dataframe with standardized column names: \code{SNP}, \code{beta},
\code{se}, \code{pval}, \code{effect_allele}, \code{other_allele}, \code{eaf}.
}
\description{
This f(x) loads GWAS summary statistics from a file and standardizes
the column names.
}
\examples{
\dontrun{
#e.g. with custom column mapping
col_map <- list(SNP = "rsid", beta "b", se = "standard_error",
                pval = "p", effect_allele = "a", other_allele = "a2",
                eaf = "freq")
gwas_data <- load_gwas_data("path/to/gwas.txt", col_map)
}

}
