% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_instruments}
\alias{select_instruments}
\title{Selecting instruments for MR}
\usage{
select_instruments(
  data,
  pval_threshold = 5e-08,
  clump = TRUE,
  r2 = 0.001,
  kb = 10000
)
}
\arguments{
\item{data}{\code{dataframe} containing exposureGWAS summary statistics
with standardized columts.}

\item{pval_threshold}{\code{numeric} p-value threshold for instrument
selection (defaul: 5e-8).}

\item{clump}{\code{logical} indicating whether to perform LD clumping
(default: 0.001).}

\item{r2}{\code{numeric} LD r-squared threshold for clumping (default: TRUE).}

\item{kb}{\code{numeric} clumping window in kilo-bases (default: 10000).}
}
\value{
a datafrae of selected instruments.
}
\description{
This f(x) selects genetic instruments for exposure GWAS data
based on a p-value threshold and optionally performs LD clumping.
}
\examples{
\dontrun{
# Select instruments with clumping
instruments <- select_instrument(exposure_data, clump = FALSE)
}

}
